added the coursecontroller to display the courses & removed the interface for easier understanding
